move: 로드맵 리스트 컴포넌트 파일 경로 수정
