+++ src/components/Roadmap/RoadmapList/RecruitRoadmap/RecruitRoadmapList/index.tsx
@@ -2,9 +2,9 @@ import { useEffect } from 'react';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 import { useGetRoadmaps } from '@/api/hooks/roadmap';
-import GroupCard from '@/components/Roadmap/GroupCard';
-import * as Styled from '@/components/Roadmap/RecruitRoadmap/RecruitRoadmapList/style';
-import TilyCard from '@/components/Roadmap/TilyCard';
+import GroupCard from '@/components/Roadmap/RoadmapList/GroupCard';
+import * as Styled from '@/components/Roadmap/RoadmapList/RecruitRoadmap/RecruitRoadmapList/style';
+import TilyCard from '@/components/Roadmap/RoadmapList/TilyCard';
 import ConditionalRender from '@/components/common/ConditionalRender';
 import CustomSuspense from '@/components/common/CustomSuspense';
 import { useIntersectionObserver } from '@/hooks/useInterSectionObserver';
