+++ src/components/Roadmap/RoadmapList/RoadmapCodeModal/index.tsx
@@ -1,6 +1,6 @@
 import { Controller, type SubmitHandler, useForm } from 'react-hook-form';
 import { usePostRoadmapsGroupsParticipate } from '@/api/hooks/roadmap';
-import * as Styled from '@/components/Roadmap/RoadmapCodeModal/style';
+import * as Styled from '@/components/Roadmap/RoadmapList/RoadmapCodeModal/style';
 import Button from '@/components/common/Button';
 import Input from '@/components/common/Input';
 import Modal, { type ModalProps } from '@/components/common/Modal';
