+++ src/pages/roadmap/[roadmapId]/index.tsx
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+const RoadmapDetail = () => {
+  return <RoadmapDetailPage></RoadmapDetailPage>;
+};
+
+export default RoadmapDetail;
+
+const RoadmapDetailPage = styled.main`
+  display: flex;
+  flex-direction: column;
+  max-width: 930px;
+  margin: 0 auto;
+`;
